Added global variable hide
The new technique loads the globals in a hide object calculation.
Instead of using conditional formatting. This allows you to use the
calculation engine to assist entering the let statement.

To ensure nothing gets displayed incorrectly, the button bars on the
layout are grouped together and set to hide if the layout name doesn’t
match a global layout name. The global layout name is set in the hide
object calc from above, so it won’t accidentally display if the globals
are not loaded on that layout. It uses a global called

Here is the "Hide object when” condition that goes on the grouped
button bars: Get ( LayoutName )  ≠ $$SORT.LAYOUTNAME

Here is the global variable setter that goes in the "Hide object when”
condition of the Global setter object.

// Page specific marker to ensure globals are loaded
$$SORT.LAYOUTNAME = Get (LayoutName);